gene_split_merge/requirements.txt
# Core dependencies
biopython>=1.79
pandas>=1.3.0

# Optional development dependencies
# Install with: pip install -r requirements.txt -r requirements-dev.txt
